Iteration K4: Adding a Locale Switcher
Add a locale selector to the application layout that allows the user to
choose from available locales and switches the locale based on the
locale selected. Added locale switcher to store catalog system test.
Added javascript pack and tag for locale switcher.

NOTE: The implementation of this should probably be reworked. The locale
switcher always redirects to the store index path, which would not be
the best user experience if you are on any other page of the app.