Iteration L2: Storing Support Requests from mail
Iteration L2: Storing Support Requests from Our Mailbox. Add
SupportRequest model to capture support requests when an email is
received by the email defined for the SupportMailbox. Also added tests
to SupportMailboxTest to ensure that the SupportRequest is captured
correctly whether there is an associated Order or not.

Also fixed bug in the Agile Web Development with Rails 6 book. The book
did not define the has_many :support_requests end of the relationship.
Logged errata on DevTalk.