Fix skipped tests and functionality errors
Fixed a couple of skipped system tests that weren't implemented yet.
Played around with the React component to try and pass dynamic data to
it, but this is not a very good use case because the edit action of an
order won't have access to any of the pay type attributes. Fixed and
added some additional translations. There is still some random locale
switching because not all links in the app have been replaced to pass
locale as part of the link.